Update Ex_4_17.java
fixed types

+++ 4_17-24/src/main/java/org/example/Ex_4_17.java
@@ -7,16 +7,16 @@ public class Ex_4_17
     public static void main(String[] args)
     {
         Scanner input = new Scanner(System.in);
-        int miles;
-        int gallons;
-        int totalMiles = 0;
-        int totalGallons = 0;
-        float MilesPerGallon;
+        double miles;
+        double gallons;
+        double totalMiles = 0;
+        double totalGallons = 0;
+        double MilesPerGallon;
 
         while(true)
         {
             System.out.print("Enter driven miles, or -1 to end program: ");
-            miles = input.nextInt();
+            miles = input.nextDouble();
 
             if(miles == -1)
             {
@@ -24,15 +24,15 @@ public class Ex_4_17
             }
 
             System.out.print("Enter gallons used: ");
-            gallons = input.nextInt();
+            gallons = input.nextDouble();
 
-            MilesPerGallon = (float)(miles/gallons);
+            MilesPerGallon = (miles/gallons);
             System.out.printf("Miles per gallon for this trip: %.2f%n", MilesPerGallon);
 
             totalMiles += miles;
             totalGallons += gallons;
 
-            System.out.printf("Miles per gallon for all trips: %.2f%n%n", (float)(totalMiles/totalGallons));
+            System.out.printf("Miles per gallon for all trips: %.2f%n%n", (totalMiles/totalGallons));
         }
     }
 }
